2023/backend/module_05/nest-clean/src/domain/forum/application/use-cases/fetch-question-answers/fetch-question-answers-use-case.ts
import { Either, right } from '@/core/either'
import { AnswersRepository } from '@/domain/forum/application/repositories'
import { Answer } from '@/domain/forum/enterprise/entities'

type Request = {
  questionId: string
  page: number
}

type Response = Either<
  null,
  {
    answers: Answer[]
  }
>

class FetchQuestionAnswersUseCase {
  constructor(private answersRepository: AnswersRepository) {}

  public async execute(request: Request): Promise<Response> {
    const { questionId, page } = request

    const answers = await this.answersRepository.findManyByQuestionId(
      questionId,
      {
        page,
      },
    )

    return right({
      answers,
    })
  }
}

export { FetchQuestionAnswersUseCase }
